# apis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/files/token": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    },
                    {
                        "RecordAuth": []
                    }
                ],
                "description": "Генерирует токен для доступа к файлу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Генерировать токен файла",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/files/{collection}/{recordId}/{filename}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    },
                    {
                        "RecordAuth": []
                    }
                ],
                "description": "Загружает файл",
                "tags": [
                    "Files"
                ],
                "summary": "Загрузить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор записи",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Имя файла",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Токен доступа к файлу",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Размер эскиза (если применимо)",
                        "name": "thumb",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл загружен"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Not exists.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "head": {
                "security": [
                    {
                        "AdminAuth": []
                    },
                    {
                        "RecordAuth": []
                    }
                ],
                "description": "Загружает файл",
                "tags": [
                    "Files"
                ],
                "summary": "Загрузить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор записи",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Имя файла",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Токен доступа к файлу",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Размер эскиза (если применимо)",
                        "name": "thumb",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл загружен"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Not exists.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}